Revert "Data logs"
