+++ src/teemtoo/DataManager.java
@@ -1,14 +1,10 @@
-package teemtoo.logic;
+package teemtoo;
 
 import javafx.beans.binding.Bindings;
 import javafx.beans.property.ObjectProperty;
 import javafx.beans.property.SimpleObjectProperty;
 import javafx.beans.value.ObservableBooleanValue;
 import teemtoo.event.Event;
-import teemtoo.sensor.Pedometer;
-import teemtoo.sensor.PulseReader;
-import teemtoo.sensor.Sensor;
-import teemtoo.tracker.*;
 
 import java.util.Arrays;
 import java.util.List;
@@ -16,8 +12,6 @@ import java.util.List;
 /**
  * @since 11/17/2017
  */
-
-//Controller of Trackers
 public final class DataManager {
 
     private static DataManager instance = new DataManager();
@@ -46,10 +40,6 @@ public final class DataManager {
 
     public Tracker getCurrentTracker() {
         return current.get();
-    }
-
-    public DataLog getCurrentLog() {
-        return current.get().getLog();
     }
 
     public List<Sensor> getSensors() {
