+++ src/teemtoo/Sensor.java
@@ -1,7 +1,6 @@
-package teemtoo.sensor;
+package teemtoo;
 
 import teemtoo.event.Event;
-import teemtoo.logic.DataManager;
 
 /**
  * @since 11/19/2017
